+++ src/main/java/de/maxhenkel/camerautils/KeyEvents.java
@@ -1,11 +1,14 @@
 package de.maxhenkel.camerautils;
 
+import com.mojang.blaze3d.platform.InputConstants;
+import com.mojang.blaze3d.platform.Window;
 import de.maxhenkel.camerautils.config.ClientConfig;
 import de.maxhenkel.camerautils.gui.*;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.minecraft.client.CameraType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
+import org.lwjgl.glfw.GLFW;
 
 public class KeyEvents {
 
@@ -203,11 +206,19 @@ public class KeyEvents {
     }
 
     public static boolean isModifierDown() {
-        if (CameraUtils.CLIENT_CONFIG.modifierKey.get().equals(ClientConfig.ModifierKey.CTRL)) {
-            return mc.hasControlDown();
-        } else {
-            return mc.hasAltDown();
-        }
+        Window window = mc.getWindow();
+        ClientConfig.ModifierKey modifierKey = CameraUtils.CLIENT_CONFIG.modifierKey.get();
+        return switch (modifierKey) {
+            case LEFT_CTRL -> InputConstants.isKeyDown(window, GLFW.GLFW_KEY_LEFT_CONTROL);
+            case RIGHT_CTRL -> InputConstants.isKeyDown(window, GLFW.GLFW_KEY_RIGHT_CONTROL);
+            case CTRL -> mc.hasControlDown();
+            case LEFT_ALT -> InputConstants.isKeyDown(window, GLFW.GLFW_KEY_LEFT_ALT);
+            case RIGHT_ALT -> InputConstants.isKeyDown(window, GLFW.GLFW_KEY_RIGHT_ALT);
+            case ALT -> mc.hasAltDown();
+            case LEFT_SHIFT -> InputConstants.isKeyDown(window, GLFW.GLFW_KEY_LEFT_SHIFT);
+            case RIGHT_SHIFT -> InputConstants.isKeyDown(window, GLFW.GLFW_KEY_RIGHT_SHIFT);
+            case SHIFT -> mc.hasShiftDown();
+        };
     }
 
 }